created Button class, adjusted Vals, improvement to CustomBox
Button class was created to be more consistent with the current
development style; the CustomButtonFactory class was over-engineered and
made the entire program less readable. It will be removed soon.
Misc adjustments to Vals.
CustomBox was setting stroke thickness using an int when it could accept
a double. Fixed.

+++ src/main/java/code/frontend/foundation/CustomButtonFactory.java
@@ -15,7 +15,7 @@ import javafx.util.Duration;
 public abstract class CustomButtonFactory
 {
     private Color feedbackColour = Vals.Colour.FEEDBACK;
-    private int btnBorderThickness = Vals.GraphicalUI.DRAW_THICKNESS;
+    private double btnBorderThickness = Vals.GraphicalUI.DRAW_THICKNESS;
     private Font labelFont = new Font(Vals.FontTools.FONT_FAM, 13);
     private Pane hoverPane = null;
     private Pane clickPane = null;
